requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
flask-socketio==5.5.1
eventlet==0.40.3
python-dotenv==1.2.1
werkzeug==3.1.3
flask-cors==6.0.1
psutil==7.1.3